// peso_lucky_admin-main/src/pages/RiskManage/PrizePool/data.d.ts
export interface Params {}
export interface WinningListItem {
  current: number;
  pageSize: number;
  jackpot_name: string;
  game_name: string;
  game_type: number;
  start_time: number;
  end_time: number;
  mer_comm_ratio: number;
  minor_amount: number;
  grand_amount: number;
  major_amount: number;
  game_id: number;
}